backend/requirements.txt
flask
flask-cors
ultralytics==8.1.23
Pillow
gunicorn